AuthService-Refactoring: Refactored package names of the project.
