+++ src/main/java/com/umssonline/auth/models/dto/Credentials.java
@@ -1,4 +1,4 @@
-package com.umssonline.auth.controllers.dto;
+package com.umssonline.auth.models.dto;
 
 public class Credentials {
